requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
apiview==1.3.25
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-import-export==3.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
MarkupPy==1.14
multidict==6.0.4
odfpy==1.4.1
openpyxl==3.1.2
psycopg2==2.9.1
psycopg2-binary==2.9.5
pycparser==2.21
pyTelegramBotAPI==4.11.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
sqlparse==0.4.4
tablib==3.4.0
urllib3==2.0.2
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
